A limited amount of function testing is possible. This will be removed
A limited amount of function testing is possible. This will be removed
in the next check-in, since it's not flexible enough to warrant
inclusion in the library; it's most likely better to make a private
function-tester. Currently testing "(const (const not) :: Double -> Int
-> Bool -> Bool) !==! const (const not)" takes ages since around one
million tests are performed. Furthermore only top-level functions are
allowed:
> [not] !==! [const True]
*** Exception: The generic versions of (==!) and friends do not accept non-bottom functions.